azure_requirement.txt
adal==1.2.7
azure-common==1.1.28
azure-core==1.29.4
azure-identity==1.14.0
azure-mgmt-core==1.4.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-resource==23.0.1
azure-mgmt-resourcegraph==8.0.0
azure-mgmt-subscription==3.1.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
idna==3.4
isodate==0.6.1
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
oauthlib==3.2.2
pathspec==0.11.2
platformdirs==3.11.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pywin32==306
setuptools==58.1.0